refactor: split GM runtime into modular modules (constants, state_defaults, state_ensure, turn_utils) and update imports
