+++ ROguelike 3_10_2025 filut jarjestyksessa luultavasti vikaa/core/gm/runtime.js
@@ -15,458 +15,33 @@
 
 import { attachGlobal } from "../../utils/global.js";
 
+import {
+  SCHEMA_VERSION,
+  MAX_DEBUG_EVENTS,
+  MAX_INTENT_HISTORY,
+  ENTRANCE_INTENT_COOLDOWN_TURNS,
+  HINT_INTENT_COOLDOWN_TURNS,
+  BOREDOM_SMOOTHING_ALPHA,
+  MOOD_TRANSIENT_DECAY_PER_TURN,
+  GUARD_FINE_HEAT_TURNS,
+} from "./runtime/constants.js";
+
+import { createDefaultMood, createDefaultState } from "./runtime/state_defaults.js";
+
+import {
+  ensureStats,
+  ensureTraitsAndMechanics,
+  ensureFactionEvents,
+  getMechanicKnowledge,
+  normalizeMood,
+  labelMood,
+} from "./runtime/state_ensure.js";
+
+import { localClamp, normalizeTurn, getCurrentTurn } from "./runtime/turn_utils.js";
+
 let _state = null;
-const SCHEMA_VERSION = 6;
-const MAX_DEBUG_EVENTS = 50;
-const MAX_INTENT_HISTORY = 20;
-const ACTION_COOLDOWN_TURNS = 80;
-// Entrance flavor can be a bit more frequent than mechanics hints, but we still keep it rare.
-const ENTRANCE_INTENT_COOLDOWN_TURNS = 60;
-// Hints stay on the original, rarer cadence.
-const HINT_INTENT_COOLDOWN_TURNS = ACTION_COOLDOWN_TURNS;
-const MOOD_DECAY_PER_TURN = 0.98;
-const MOOD_BOREDOM_HIGH = 0.8;
-const MOOD_BOREDOM_LOW = 0.2;
-const MOOD_RECENT_INTERESTING_TURNS = 30;
-const MOOD_NUDGE_DELTA = 0.01;
-const MOOD_NUDGE_DELTA_SMALL = 0.005;
-const BOREDOM_SMOOTHING_ALPHA = 0.15; // how fast boredom reacts to changes
-const MOOD_TRANSIENT_DECAY_PER_TURN = 0.9; // transients decay faster than base mood
 
-// Mechanic knowledge thresholds; see getMechanicKnowledge.
-const MECH_RECENT_TURNS = 200;
-const MECH_DISINTEREST_DISMISS = 3;
-const MECH_DISINTEREST_AGE = 600;
-// Guard fine "heat" forget window; same scale as trait forgetting.
-const GUARD_FINE_HEAT_TURNS = 300;
 
-function createDefaultStats() {
-  return {
-    totalTurns: 0,
-    // turns spent in each mode/scope, e.g. { world: 123, town: 45, dungeon: 67, ... }
-    modeTurns: {},
-    // number of times we entered each mode/scope via events, e.g. { world: 10, town: 5, ... }
-    modeEntries: {},
-    // encounter counts
-    encounterStarts: 0,
-    encounterCompletions: 0,
-  };
-}
-
-function createDefaultTraits() {
-  return {
-    trollSlayer: { seen: 0, positive: 0, negative: 0, lastUpdatedTurn: null },
-    townProtector: { seen: 0, positive: 0, negative: 0, lastUpdatedTurn: null },
-    caravanAlly: { seen: 0, positive: 0, negative: 0, lastUpdatedTurn: null },
-  };
-}
-
-function createDefaultMechanics() {
-  return {
-    fishing: { seen: 0, tried: 0, success: 0, failure: 0, dismiss: 0, firstSeenTurn: null, lastUsedTurn: null },
-    lockpicking: { seen: 0, tried: 0, success: 0, failure: 0, dismiss: 0, firstSeenTurn: null, lastUsedTurn: null },
-    questBoard: { seen: 0, tried: 0, success: 0, failure: 0, dismiss: 0, firstSeenTurn: null, lastUsedTurn: null },
-    followers: { seen: 0, tried: 0, success: 0, failure: 0, dismiss: 0, firstSeenTurn: null, lastUsedTurn: null },
-  };
-}
-
-function createDefaultMood() {
-  return {
-    primary: "neutral",
-    valence: 0.0,
-    arousal: 0.0,
-    lastUpdatedTurn: null,
-    baselineValence: 0.0,
-    baselineArousal: 0.0,
-    transientValence: 0.0,
-    transientArousal: 0.0,
-  };
-}
-
-function createDefaultState() {
-  return {
-    schemaVersion: SCHEMA_VERSION,
-    enabled: true,
-    mood: createDefaultMood(),
-    boredom: {
-      level: 0.0,
-      turnsSinceLastInterestingEvent: 0,
-      lastInterestingEvent: null,
-    },
-    storyFlags: {
-      factionEvents: {},
-    },
-    debug: {
-      enabled: false,
-      logTicks: false,
-      logEvents: false,
-      lastTickTurn: -1,
-      lastEvent: null,
-      lastEvents: [],
-      counters: {
-        ticks: 0,
-        events: 0,
-        interestingEvents: 0,
-      },
-      lastIntent: null,
-      intentHistory: [],
-    },
-    stats: createDefaultStats(),
-    traits: createDefaultTraits(),
-    mechanics: createDefaultMechanics(),
-    families: {},
-    factions: {},
-    lastMode: "world",
-    // Legacy/global last GM action turn; keep for back-compat and debugging.
-    lastActionTurn: -1,
-    // Separate cooldown tracking for entrance flavor vs mechanic hints.
-    lastEntranceIntentTurn: -1,
-    lastHintIntentTurn: -1,
-    lastHintIntentTownEntry: -1,
-  };
-}
-
-function localClamp(v, lo, hi) {
-  if (v < lo) return lo;
-  if (v > hi) return hi;
-  return v;
-}
-
-function ensureStats(gm) {
-  if (!gm || typeof gm !== "object") {
-    return createDefaultStats();
-  }
-
-  let stats = gm.stats;
-  if (!stats || typeof stats !== "object") {
-    stats = createDefaultStats();
-    gm.stats = stats;
-  }
-
-  // Normalize fields to simple integer counters and plain-object maps.
-  stats.totalTurns = stats.totalTurns | 0;
-  if (!stats.modeTurns || typeof stats.modeTurns !== "object") {
-    stats.modeTurns = {};
-  }
-  if (!stats.modeEntries || typeof stats.modeEntries !== "object") {
-    stats.modeEntries = {};
-  }
-  stats.encounterStarts = stats.encounterStarts | 0;
-  stats.encounterCompletions = stats.encounterCompletions | 0;
-
-  return stats;
-}
-
-function ensureTraitsAndMechanics(gm) {
-  if (!gm || typeof gm !== "object") return;
-
-  if (!gm.storyFlags || typeof gm.storyFlags !== "object") {
-    gm.storyFlags = {};
-  }
-
-  // Traits
-  let traits = gm.traits;
-  if (!traits || typeof traits !== "object") {
-    traits = createDefaultTraits();
-    gm.traits = traits;
-  }
-
-  function normalizeNonNegative(obj, key) {
-    let v = obj[key] | 0;
-    if (v < 0) v = 0;
-    obj[key] = v;
-  }
-
-  function normalizeTurnField(obj, key) {
-    const raw = obj[key];
-    if (raw == null) {
-      obj[key] = null;
-      return;
-    }
-    let v = typeof raw === "number" ? (raw | 0) : NaN;
-    if (!Number.isFinite(v) || v < 0) {
-      obj[key] = null;
-    } else {
-      obj[key] = v;
-    }
-  }
-
-  function ensureTrait(key) {
-    if (!traits[key] || typeof traits[key] !== "object") {
-      traits[key] = {
-        seen: 0,
-        positive: 0,
-        negative: 0,
-        lastUpdatedTurn: null,
-      };
-    }
-    const t = traits[key];
-    normalizeNonNegative(t, "seen");
-    normalizeNonNegative(t, "positive");
-    normalizeNonNegative(t, "negative");
-    normalizeTurnField(t, "lastUpdatedTurn");
-  }
-
-  ensureTrait("trollSlayer");
-  ensureTrait("townProtector");
-  ensureTrait("caravanAlly");
-
-  // Mechanics
-  let mechanics = gm.mechanics;
-  if (!mechanics || typeof mechanics !== "object") {
-    mechanics = createDefaultMechanics();
-    gm.mechanics = mechanics;
-  }
-
-  function ensureMechanic(key) {
-    if (!mechanics[key] || typeof mechanics[key] !== "object") {
-      mechanics[key] = {
-        seen: 0,
-        tried: 0,
-        success: 0,
-        failure: 0,
-        dismiss: 0,
-        firstSeenTurn: null,
-        lastUsedTurn: null,
-      };
-    }
-    const m = mechanics[key];
-    normalizeNonNegative(m, "seen");
-    normalizeNonNegative(m, "tried");
-    normalizeNonNegative(m, "success");
-    normalizeNonNegative(m, "failure");
-    normalizeNonNegative(m, "dismiss");
-    normalizeTurnField(m, "firstSeenTurn");
-    normalizeTurnField(m, "lastUsedTurn");
-  }
-
-  ensureMechanic("fishing");
-  ensureMechanic("lockpicking");
-  ensureMechanic("questBoard");
-  ensureMechanic("followers");
-
-  // Families: ensure container exists and normalize existing entries.
-  let families = gm.families;
-  if (!families || typeof families !== "object") {
-    families = {};
-    gm.families = families;
-  } else {
-    for (const key in families) {
-      if (Object.prototype.hasOwnProperty.call(families, key)) {
-        normalizeFamilyTrait(families[key]);
-      }
-    }
-  }
-
-  // Factions: ensure container exists and normalize existing entries.
-  let factions = gm.factions;
-  if (!factions || typeof factions !== "object" || Array.isArray(factions)) {
-    factions = {};
-    gm.factions = factions;
-  } else {
-    for (const key in factions) {
-      if (Object.prototype.hasOwnProperty.call(factions, key)) {
-        normalizeFamilyTrait(factions[key]);
-      }
-    }
-  }
-
-  // Debug ring buffers
-  if (gm.debug && typeof gm.debug === "object") {
-    const dbg = gm.debug;
-    if (!Array.isArray(dbg.lastEvents)) {
-      dbg.lastEvents = [];
-    }
-    if (dbg.lastEvents.length > MAX_DEBUG_EVENTS) {
-      dbg.lastEvents.length = MAX_DEBUG_EVENTS;
-    }
-    if (!Array.isArray(dbg.intentHistory)) {
-      dbg.intentHistory = [];
-    }
-    if (dbg.intentHistory.length > MAX_INTENT_HISTORY) {
-      dbg.intentHistory.length = MAX_INTENT_HISTORY;
-    }
-  }
-}
-
-function ensureFactionEvents(gm) {
-  if (!gm || typeof gm !== "object") return;
-
-  if (!gm.storyFlags || typeof gm.storyFlags !== "object") {
-    gm.storyFlags = {};
-  }
-
-  let factionEvents = gm.storyFlags.factionEvents;
-  if (!factionEvents || typeof factionEvents !== "object" || Array.isArray(factionEvents)) {
-    factionEvents = {};
-    gm.storyFlags.factionEvents = factionEvents;
-  }
-
-  function normalizeTurnField(value) {
-    if (typeof value !== "number") return 0;
-    let v = value | 0;
-    if (v < 0) v = 0;
-    return v;
-  }
-
-  function ensureSlot(key) {
-    let slot = factionEvents[key];
-    if (!slot || typeof slot !== "object") {
-      slot = {};
-      factionEvents[key] = slot;
-    }
-
-    let status = typeof slot.status === "string" ? slot.status : "none";
-    if (status !== "none" && status !== "scheduled" && status !== "consumed") {
-      status = "none";
-    }
-    slot.status = status;
-
-    const earliest = normalizeTurnField(slot.earliestTurn);
-    let latest = normalizeTurnField(slot.latestTurn);
-
-    if (latest !== 0 && latest < earliest) {
-      latest = earliest;
-    }
-
-    slot.earliestTurn = earliest;
-    slot.latestTurn = latest;
-  }
-
-  ensureSlot("banditBounty");
-  ensureSlot("guardFine");
-  ensureSlot("trollHunt");
-}
-
-// Returns one of: "unseen", "seenNotTried", "triedRecently", "triedLongAgo", "disinterested".
-// Derived from counts and lastUsedTurn; does not mutate the mechanic object.
-function getMechanicKnowledge(m, currentTurn) {
-  if (!m || typeof m !== "object") return "unseen";
-
-  const seen = m.seen | 0;
-  const tried = m.tried | 0;
-  const dismiss = m.dismiss | 0;
-
-  let turn = typeof currentTurn === "number" ? (currentTurn | 0) : 0;
-  if (turn < 0) turn = 0;
-
-  let normalizedLastUsed = null;
-  const rawLastUsed = m.lastUsedTurn;
-  if (rawLastUsed != null) {
-    const num = typeof rawLastUsed === "number" ? (rawLastUsed | 0) : NaN;
-    if (Number.isFinite(num) && num >= 0) {
-      normalizedLastUsed = num;
-    }
-  }
-
-  let age = null;
-  if (normalizedLastUsed != null) {
-    age = turn - normalizedLastUsed;
-    if (age < 0) age = 0;
-  }
-
-  if (seen <= 0 && tried <= 0) {
-    return "unseen";
-  }
-
-  if (tried === 0) {
-    if (dismiss >= MECH_DISINTEREST_DISMISS) {
-      return "disinterested";
-    }
-    return "seenNotTried";
-  }
-
-  // tried > 0
-  if (normalizedLastUsed == null) {
-    age = MECH_DISINTEREST_AGE + 1;
-  }
-
-  if (dismiss >= MECH_DISINTEREST_DISMISS && age != null && age > MECH_RECENT_TURNS) {
-    return "disinterested";
-  }
-
-  if (age != null && age <= MECH_RECENT_TURNS) {
-    return "triedRecently";
-  }
-
-  if (age != null && age > MECH_DISINTEREST_AGE) {
-    return "disinterested";
-  }
-
-  return "triedLongAgo";
-}
-
-function normalizeFamilyTrait(family) {
-  if (!family || typeof family !== "object") return;
-  family.seen = (family.seen | 0);
-  if (family.seen < 0) family.seen = 0;
-  family.positive = (family.positive | 0);
-  if (family.positive < 0) family.positive = 0;
-  family.negative = (family.negative | 0);
-  if (family.negative < 0) family.negative = 0;
-  const rawTurn = family.lastUpdatedTurn;
-  if (rawTurn == null) {
-    family.lastUpdatedTurn = null;
-  } else {
-    let t = typeof rawTurn === "number" ? (rawTurn | 0) : NaN;
-    if (!Number.isFinite(t) || t < 0) family.lastUpdatedTurn = null;
-    else family.lastUpdatedTurn = t;
-  }
-}
-
-function normalizeMood(gm) {
-  if (!gm || typeof gm !== "object") return;
-  let mood = gm.mood;
-  if (!mood || typeof mood !== "object") {
-    mood = createDefaultMood();
-    gm.mood = mood;
-  }
-
-  // Ensure extended mood fields exist with safe defaults.
-  if (!Number.isFinite(mood.baselineValence)) mood.baselineValence = 0;
-  if (!Number.isFinite(mood.baselineArousal)) mood.baselineArousal = 0;
-  if (!Number.isFinite(mood.transientValence)) mood.transientValence = 0;
-  if (!Number.isFinite(mood.transientArousal)) mood.transientArousal = 0;
-
-  let v = Number(mood.valence);
-  let a = Number(mood.arousal);
-  if (!Number.isFinite(v)) v = 0;
-  if (!Number.isFinite(a)) a = 0;
-
-  v = localClamp(v, -1, 1);
-  a = localClamp(a, 0, 1);
-
-  mood.valence = v;
-  mood.arousal = a;
-  if (typeof mood.primary !== "string") {
-    mood.primary = "neutral";
-  }
-}
-
-function labelMood(gm, turn) {
-  if (!gm || typeof gm !== "object") return;
-  const mood = gm.mood;
-  if (!mood || typeof mood !== "object") return;
-
-  const v = typeof mood.valence === "number" && Number.isFinite(mood.valence) ? mood.valence : 0;
-  const a = typeof mood.arousal === "number" && Number.isFinite(mood.arousal) ? mood.arousal : 0;
-
-  let primary = "neutral";
-  if (a < 0.2) {
-    primary = "calm";
-  } else if (a < 0.6) {
-    if (v > 0.2) primary = "curious";
-    else if (v < -0.2) primary = "bored";
-    else primary = "neutral";
-  } else {
-    if (v > 0.2) primary = "playful";
-    else if (v < -0.2) primary = "stern";
-    else primary = "restless";
-  }
-
-  mood.primary = primary;
-  mood.lastUpdatedTurn = normalizeTurn(turn);
-}
 
 function addMoodImpulse(gm, baseValence, baseArousal) {
   if (!gm || typeof gm !== "object") return;
@@ -1049,21 +624,7 @@ function maybeScheduleFactionEvents(ctx, gm, turn) {
   }
 }
 
-function normalizeTurn(turn) {
-  let t = typeof turn === "number" ? (turn | 0) : 0;
-  if (t < 0) t = 0;
-  return t;
-}
 
-function getCurrentTurn(ctx, gm) {
-  if (ctx && ctx.time && typeof ctx.time.turnCounter === "number") {
-    return ctx.time.turnCounter | 0;
-  }
-  if (gm && gm.debug && typeof gm.debug.lastTickTurn === "number") {
-    return gm.debug.lastTickTurn | 0;
-  }
-  return 0;
-}
 
 function extractFamilyKeyFromTags(rawTags) {
   if (!Array.isArray(rawTags) || rawTags.length === 0) return null;
